+++ modules/bag/apps/bag_recorder.cpp
@@ -11,8 +11,8 @@
 #include "hephaestus/bag/writer.h"
 #include "hephaestus/bag/zenoh_recorder.h"
 #include "hephaestus/cli/program_options.h"
-#include "hephaestus/ipc/program_options.h"
 #include "hephaestus/ipc/topic_filter.h"
+#include "hephaestus/ipc/zenoh/program_options.h"
 #include "hephaestus/ipc/zenoh/session.h"
 #include "hephaestus/utils/signal_handler.h"
 #include "hephaestus/utils/stack_trace.h"
@@ -22,10 +22,10 @@ auto main(int argc, const char* argv[]) -> int {
 
   try {
     auto desc = heph::cli::ProgramDescription("Record a bag from zenoh topics");
-    heph::ipc::appendIPCProgramOption(desc);
+    heph::ipc::zenoh::appendProgramOption(desc);
     desc.defineOption<std::string>("output_bag", 'o', "output file where to write the bag");
     const auto args = std::move(desc).parse(argc, argv);
-    auto [config, topic] = heph::ipc::parseIPCProgramOptions(args);
+    auto [config, topic] = heph::ipc::zenoh::parseProgramOptions(args);
     auto output_file = args.getOption<std::string>("output_bag");
 
     heph::bag::ZenohRecorderParams params{
