+++ modules/examples/examples/zenoh_action_server_client.cpp
@@ -14,9 +14,11 @@
 #include <fmt/core.h>
 #include <magic_enum.hpp>
 
+#include "hephaestus/cli/program_options.h"
 #include "hephaestus/examples/types/sample.h"
 #include "hephaestus/examples/types_protobuf/sample.h"  // NOLINT(misc-include-cleaner)
 #include "hephaestus/ipc/zenoh/action_server/action_server.h"
+#include "hephaestus/ipc/zenoh/program_options.h"
 #include "hephaestus/ipc/zenoh/session.h"
 #include "hephaestus/utils/signal_handler.h"
 #include "hephaestus/utils/stack_trace.h"
@@ -26,10 +28,11 @@ auto main(int argc, const char* argv[]) -> int {
   const heph::utils::StackTrace stack_trace;
 
   try {
-    auto desc = getProgramDescription("Binary service client example", ExampleType::ACTION_SERVER);
+    auto desc = heph::cli::ProgramDescription("Action server client example");
+    heph::ipc::zenoh::appendProgramOption(desc, getDefaultTopic(ExampleType::ACTION_SERVER));
     const auto args = std::move(desc).parse(argc, argv);
+    auto [session_config, topic_config] = heph::ipc::zenoh::parseProgramOptions(args);
 
-    auto [session_config, topic_config] = parseArgs(args);
     auto session = heph::ipc::zenoh::createSession(std::move(session_config));
 
     heph::utils::TerminationBlocker::registerInterruptCallback([session = std::ref(*session), &topic_config] {
