+++ modules/examples/examples/zenoh_string_service_client.cpp
@@ -15,7 +15,9 @@
 #include <fmt/chrono.h>  // NOLINT(misc-include-cleaner)
 #include <fmt/core.h>
 
+#include "hephaestus/cli/program_options.h"
 #include "hephaestus/examples/types_protobuf/pose.h"  // NOLINT(misc-include-cleaner)
+#include "hephaestus/ipc/zenoh/program_options.h"
 #include "hephaestus/ipc/zenoh/service.h"
 #include "hephaestus/ipc/zenoh/session.h"
 #include "hephaestus/utils/stack_trace.h"
@@ -25,10 +27,11 @@ auto main(int argc, const char* argv[]) -> int {
   const heph::utils::StackTrace stack_trace;
 
   try {
-    auto desc = getProgramDescription("String service client example", ExampleType::SERVICE);
+    auto desc = heph::cli::ProgramDescription("String service client example");
+    heph::ipc::zenoh::appendProgramOption(desc, getDefaultTopic(ExampleType::SERVICE));
     const auto args = std::move(desc).parse(argc, argv);
+    auto [session_config, topic_config] = heph::ipc::zenoh::parseProgramOptions(args);
 
-    auto [session_config, topic_config] = parseArgs(args);
     auto session = heph::ipc::zenoh::createSession(std::move(session_config));
 
     static constexpr auto K_TIMEOUT = std::chrono::seconds(10);
