+++ modules/examples/examples/zenoh_string_service_server.cpp
@@ -12,6 +12,8 @@
 #include <absl/log/log.h>
 #include <fmt/core.h>
 
+#include "hephaestus/cli/program_options.h"
+#include "hephaestus/ipc/zenoh/program_options.h"
 #include "hephaestus/ipc/zenoh/service.h"
 #include "hephaestus/ipc/zenoh/session.h"
 #include "hephaestus/utils/signal_handler.h"
@@ -22,10 +24,11 @@ auto main(int argc, const char* argv[]) -> int {
   const heph::utils::StackTrace stack_trace;
 
   try {
-    auto desc = getProgramDescription("String service server example", ExampleType::SERVICE);
+    auto desc = heph::cli::ProgramDescription("String service server example");
+    heph::ipc::zenoh::appendProgramOption(desc, getDefaultTopic(ExampleType::SERVICE));
     const auto args = std::move(desc).parse(argc, argv);
+    auto [session_config, topic_config] = heph::ipc::zenoh::parseProgramOptions(args);
 
-    auto [session_config, topic_config] = parseArgs(args);
     auto session = heph::ipc::zenoh::createSession(std::move(session_config));
 
     auto callback = [](const std::string& sample) {
