+++ modules/ipc/apps/zenoh_string_service.cpp
@@ -14,7 +14,7 @@
 #include <fmt/core.h>
 
 #include "hephaestus/cli/program_options.h"
-#include "hephaestus/ipc/program_options.h"
+#include "hephaestus/ipc/zenoh/program_options.h"
 #include "hephaestus/ipc/zenoh/service.h"
 #include "hephaestus/ipc/zenoh/session.h"
 #include "hephaestus/ipc/zenoh/utils.h"
@@ -27,12 +27,12 @@ auto main(int argc, const char* argv[]) -> int {
     auto desc =
         heph::cli::ProgramDescription("Simple service for std::string types for both Request and Reply. "
                                       "Don't use for services with different types.");
-    heph::ipc::appendIPCProgramOption(desc);
+    heph::ipc::zenoh::appendProgramOption(desc);
     desc.defineOption<std::string>("value", 'v', "the value to pass the query", "");
     const auto args = std::move(desc).parse(argc, argv);
     const auto value = args.getOption<std::string>("value");
 
-    auto [config, topic_config] = heph::ipc::parseIPCProgramOptions(args);
+    auto [config, topic_config] = heph::ipc::zenoh::parseProgramOptions(args);
     auto session = heph::ipc::zenoh::createSession(std::move(config));
     LOG(INFO) << fmt::format("Opening session: {}",
                              heph::ipc::zenoh::toString(session->zenoh_session.get_zid()));
