+++ modules/ipc/apps/zenoh_topic_echo.cpp
@@ -20,9 +20,10 @@
 #include <nlohmann/json_fwd.hpp>
 
 #include "hephaestus/cli/program_options.h"
-#include "hephaestus/ipc/common.h"
-#include "hephaestus/ipc/program_options.h"
+#include "hephaestus/ipc/topic.h"
 #include "hephaestus/ipc/zenoh/dynamic_subscriber.h"
+#include "hephaestus/ipc/zenoh/program_options.h"
+#include "hephaestus/ipc/zenoh/raw_subscriber.h"
 #include "hephaestus/ipc/zenoh/session.h"
 #include "hephaestus/serdes/dynamic_deserializer.h"
 #include "hephaestus/serdes/type_info.h"
@@ -55,26 +56,26 @@ void truncateLongItems(std::string& msg_json, bool noarr, size_t max_length) {
 }
 }  // namespace
 
-namespace heph::ipc::apps {
+namespace heph::ipc::zenoh::apps {
 class TopicEcho {
 public:
-  TopicEcho(zenoh::SessionPtr session, TopicConfig topic_config, bool noarr, std::size_t max_array_length)
+  TopicEcho(SessionPtr session, TopicConfig topic_config, bool noarr, std::size_t max_array_length)
     : noarr_(noarr), max_array_length_(max_array_length), topic_config_(std::move(topic_config)) {
-    zenoh::DynamicSubscriberParams params{
-      .session = std::move(session),
-      .topics_filter_params = { .include_topics_only = {},
-                                .prefix = topic_config_.name,
-                                .exclude_topics = {} },
-      .init_subscriber_cb =
-          [this](const auto& topic, const auto& type_info) {
-            (void)topic;
-            dynamic_deserializer_.registerSchema(type_info);
-          },
-      .subscriber_cb = [this](const auto& metadata, auto data,
-                              const auto& topic_info) { subscribeCallback(metadata, data, topic_info); }
-    };
+    DynamicSubscriberParams params{ .session = std::move(session),
+                                    .topics_filter_params = { .include_topics_only = {},
+                                                              .prefix = topic_config_.name,
+                                                              .exclude_topics = {} },
+                                    .init_subscriber_cb =
+                                        [this](const auto& topic, const auto& type_info) {
+                                          (void)topic;
+                                          dynamic_deserializer_.registerSchema(type_info);
+                                        },
+                                    .subscriber_cb =
+                                        [this](const auto& metadata, auto data, const auto& topic_info) {
+                                          subscribeCallback(metadata, data, topic_info);
+                                        } };
 
-    dynamic_subscriber_ = std::make_unique<zenoh::DynamicSubscriber>(std::move(params));
+    dynamic_subscriber_ = std::make_unique<DynamicSubscriber>(std::move(params));
   }
 
   [[nodiscard]] auto start() -> std::future<void> {
@@ -103,17 +104,17 @@ private:
   bool noarr_;
   std::size_t max_array_length_;
   TopicConfig topic_config_;
-  heph::serdes::DynamicDeserializer dynamic_deserializer_;
-  std::unique_ptr<zenoh::DynamicSubscriber> dynamic_subscriber_;
+  serdes::DynamicDeserializer dynamic_deserializer_;
+  std::unique_ptr<DynamicSubscriber> dynamic_subscriber_;
 };
-}  // namespace heph::ipc::apps
+}  // namespace heph::ipc::zenoh::apps
 
 auto main(int argc, const char* argv[]) -> int {
   const heph::utils::StackTrace stack_trace;
 
   try {
     auto desc = heph::cli::ProgramDescription("Echo the data from a topic to the console.");
-    heph::ipc::appendIPCProgramOption(desc);
+    heph::ipc::zenoh::appendProgramOption(desc);
 
     desc.defineFlag("noarr", "Truncate print of long arrays");
     desc.defineOption<std::size_t>(
@@ -123,7 +124,7 @@ auto main(int argc, const char* argv[]) -> int {
         DEFAULT_MAX_ARRAY_LENGTH);
     const auto args = std::move(desc).parse(argc, argv);
 
-    auto [session_config, topic_config] = heph::ipc::parseIPCProgramOptions(args);
+    auto [session_config, topic_config] = heph::ipc::zenoh::parseProgramOptions(args);
     const auto noarr = args.getOption<bool>("noarr");
     const auto max_array_length = args.getOption<std::size_t>("noarr-max-size");
 
@@ -131,7 +132,7 @@ auto main(int argc, const char* argv[]) -> int {
 
     auto session = heph::ipc::zenoh::createSession(std::move(session_config));
 
-    heph::ipc::apps::TopicEcho topic_echo{ std::move(session), topic_config, noarr, max_array_length };
+    heph::ipc::zenoh::apps::TopicEcho topic_echo{ std::move(session), topic_config, noarr, max_array_length };
     topic_echo.start().wait();
 
     heph::utils::TerminationBlocker::waitForInterrupt();
