+++ modules/ipc/include/hephaestus/ipc/zenoh/scout.h
@@ -5,7 +5,7 @@
 #pragma once
 #include <vector>
 
-#include "hephaestus/ipc/common.h"
+#include "hephaestus/ipc/zenoh/session.h"
 
 namespace heph::ipc::zenoh {
