+++ modules/ipc/include/hephaestus/ipc/zenoh/service.h
@@ -25,7 +25,7 @@
 #include <zenoh.hxx>
 #include <zenoh/api/queryable.hxx>
 
-#include "hephaestus/ipc/common.h"
+#include "hephaestus/ipc/topic.h"
 #include "hephaestus/ipc/zenoh/session.h"
 #include "hephaestus/ipc/zenoh/utils.h"
 #include "hephaestus/serdes/serdes.h"
