+++ modules/ipc/include/hephaestus/ipc/zenoh/session.h
@@ -7,10 +7,25 @@
 #include <zenoh.h>
 #include <zenoh.hxx>
 
-#include "hephaestus/ipc/common.h"
-#include "hephaestus/ipc/zenoh/utils.h"
-
 namespace heph::ipc::zenoh {
+
+enum class Mode : uint8_t { PEER = 0, CLIENT, ROUTER };
+enum class Protocol : uint8_t { ANY = 0, UDP, TCP };
+
+struct Config {
+  bool enable_shared_memory = false;  //! NOTE: With shared-memory enabled, the publisher still uses the
+                                      //! network transport layer to notify subscribers of the shared-memory
+                                      //! segment to read. Therefore, for very small messages, shared -
+                                      //! memory transport could be less efficient than using the default
+                                      //! network transport to directly carry the payload.
+  Mode mode = Mode::PEER;
+  // NOLINTNEXTLINE(readability-redundant-string-init) otherwise we need to specify in constructor
+  std::string router = "";  //! If specified connect to the given router endpoint.
+  std::size_t cache_size = 0;
+  bool qos = false;
+  bool real_time = false;
+  Protocol protocol{ Protocol::ANY };
+};
 
 struct Session {
   ::zenoh::Session zenoh_session;
@@ -18,9 +33,6 @@ struct Session {
 };
 
 using SessionPtr = std::shared_ptr<Session>;
+[[nodiscard]] auto createSession(Config config) -> SessionPtr;
 
-[[nodiscard]] inline auto createSession(Config config) -> SessionPtr {
-  auto zconfig = createZenohConfig(config);
-  return std::make_shared<Session>(::zenoh::Session::open(std::move(zconfig)), std::move(config));
-}
 }  // namespace heph::ipc::zenoh
