+++ modules/ipc/src/zenoh/action_server/action_server.cpp
@@ -10,7 +10,7 @@
 #include <absl/log/log.h>
 #include <fmt/core.h>
 
-#include "hephaestus/ipc/common.h"
+#include "hephaestus/ipc/topic.h"
 #include "hephaestus/ipc/zenoh/action_server/client_helper.h"
 #include "hephaestus/ipc/zenoh/service.h"
 #include "hephaestus/ipc/zenoh/session.h"
