+++ modules/ipc/src/zenoh/scout.cpp
@@ -15,6 +15,7 @@
 
 #include <absl/base/thread_annotations.h>
 #include <fmt/core.h>
+#include <magic_enum.hpp>
 #include <nlohmann/json.hpp>
 #include <nlohmann/json_fwd.hpp>
 #include <range/v3/range/conversion.hpp>
@@ -22,7 +23,7 @@
 #include <zenoh/api/hello.hxx>
 #include <zenoh/api/scout.hxx>
 
-#include "hephaestus/ipc/common.h"
+#include "hephaestus/ipc/topic.h"
 #include "hephaestus/ipc/zenoh/service.h"
 #include "hephaestus/ipc/zenoh/session.h"
 #include "hephaestus/ipc/zenoh/utils.h"
@@ -65,7 +66,7 @@ private:
 };
 
 [[nodiscard]] auto getRouterInfoJson(const std::string& router_id) -> std::string {
-  heph::ipc::Config config;
+  heph::ipc::zenoh::Config config;
   auto session = heph::ipc::zenoh::createSession(std::move(config));
 
   static constexpr auto ROUTER_TOPIC = "@/router/{}";
@@ -118,6 +119,7 @@ auto getListOfNodes() -> std::vector<NodeInfo> {
 }
 
 auto toString(const NodeInfo& info) -> std::string {
-  return fmt::format("[{}] ID: {}. Locators: {}", toString(info.mode), info.id, toString(info.locators));
+  return fmt::format("[{}] ID: {}. Locators: {}", magic_enum::enum_name(info.mode), info.id,
+                     toString(info.locators));
 }
 }  // namespace heph::ipc::zenoh
